Stats: SQL: Fix conquest pagination.

+++ src/sql/get-tribe-conquests-gain-count.sql
@@ -1,4 +1,4 @@
-SELECT COUNT(*)
+SELECT COUNT(*)::INT
 FROM ${worldId:name}.conquests
 WHERE new_owner_tribe_id = ${tribeId}
 AND (old_owner_tribe_id != ${tribeId} OR old_owner_tribe_id IS NULL)
